+++ src/test/java/fr/insee/genesis/domain/service/surveyunit/SurveyUnitServiceUnitTest.java
@@ -1,6 +1,7 @@
 package fr.insee.genesis.domain.service.surveyunit;
 
 import fr.insee.genesis.TestConstants;
+import fr.insee.genesis.controller.dto.SurveyUnitSimplifiedDto;
 import fr.insee.genesis.domain.model.surveyunit.DataState;
 import fr.insee.genesis.domain.model.surveyunit.Mode;
 import fr.insee.genesis.domain.model.surveyunit.SurveyUnitModel;
@@ -12,6 +13,7 @@ import fr.insee.genesis.infrastructure.mappers.SurveyUnitDocumentMapper;
 import fr.insee.genesis.infrastructure.utils.FileUtils;
 import fr.insee.genesis.stubs.ConfigStub;
 import fr.insee.genesis.stubs.SurveyUnitPersistencePortStub;
+import lombok.SneakyThrows;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -69,6 +71,166 @@ class SurveyUnitServiceUnitTest {
         Assertions.assertThat(surveyUnitModels).isNotNull().hasSize(1);
         Assertions.assertThat(surveyUnitModels.getFirst().getUsualSurveyUnitId()).isEqualTo(TestConstants.DEFAULT_SURVEY_UNIT_ID);
 
+    }
+
+    @Test
+    void get_latest_should_return_state_in_collected_variables(){
+        //GIVEN
+        String variableId = "testVar";
+        DataState state = DataState.COLLECTED;
+
+        SurveyUnitDocument surveyUnitDocument = new SurveyUnitDocument();
+        surveyUnitDocument.setState(String.valueOf(state));
+        surveyUnitDocument.setMode(String.valueOf(Mode.WEB));
+        surveyUnitDocument.setUsualSurveyUnitId(TestConstants.DEFAULT_SURVEY_UNIT_ID);
+
+        VariableDocument variableDocument = new VariableDocument();
+        variableDocument.setVarId(variableId);
+        variableDocument.setIteration(1);
+        surveyUnitDocument.setCollectedVariables(new ArrayList<>());
+        surveyUnitDocument.getCollectedVariables().add(variableDocument);
+
+        doReturn(SurveyUnitDocumentMapper.INSTANCE.listDocumentToListModel(Collections.singletonList(surveyUnitDocument)))
+                .when(surveyUnitPersistencePortStub).findInterrogationIdsByCollectionInstrumentId(any());
+        doReturn(SurveyUnitDocumentMapper.INSTANCE.listDocumentToListModel(Collections.singletonList(surveyUnitDocument)))
+                .when(surveyUnitPersistencePortStub).findByIds(any(), any());
+
+        //WHEN
+        List<SurveyUnitModel> surveyUnitModels = surveyUnitService.findLatestByIdAndByCollectionInstrumentId(
+                TestConstants.DEFAULT_INTERROGATION_ID,
+                TestConstants.DEFAULT_COLLECTION_INSTRUMENT_ID
+        );
+
+        //THEN
+        Assertions.assertThat(surveyUnitModels).isNotNull().hasSize(1);
+        Assertions.assertThat(surveyUnitModels.getFirst().getCollectedVariables().getFirst().state())
+                .isEqualTo(state);
+    }
+
+    @Test
+    void get_latest_should_return_state_in_external_variables(){
+        //GIVEN
+        String variableId = "testVar";
+        DataState state = DataState.COLLECTED;
+
+        SurveyUnitDocument surveyUnitDocument = new SurveyUnitDocument();
+        surveyUnitDocument.setState(String.valueOf(state));
+        surveyUnitDocument.setMode(String.valueOf(Mode.WEB));
+        surveyUnitDocument.setUsualSurveyUnitId(TestConstants.DEFAULT_SURVEY_UNIT_ID);
+
+        VariableDocument variableDocument = new VariableDocument();
+        variableDocument.setVarId(variableId);
+        variableDocument.setIteration(1);
+        surveyUnitDocument.setExternalVariables(new ArrayList<>());
+        surveyUnitDocument.getExternalVariables().add(variableDocument);
+
+        doReturn(SurveyUnitDocumentMapper.INSTANCE.listDocumentToListModel(Collections.singletonList(surveyUnitDocument)))
+                .when(surveyUnitPersistencePortStub).findInterrogationIdsByCollectionInstrumentId(any());
+        doReturn(SurveyUnitDocumentMapper.INSTANCE.listDocumentToListModel(Collections.singletonList(surveyUnitDocument)))
+                .when(surveyUnitPersistencePortStub).findByIds(any(), any());
+
+        //WHEN
+        List<SurveyUnitModel> surveyUnitModels = surveyUnitService.findLatestByIdAndByCollectionInstrumentId(
+                TestConstants.DEFAULT_INTERROGATION_ID,
+                TestConstants.DEFAULT_COLLECTION_INSTRUMENT_ID
+        );
+
+        //THEN
+        Assertions.assertThat(surveyUnitModels).isNotNull().hasSize(1);
+        Assertions.assertThat(surveyUnitModels.getFirst().getExternalVariables().getFirst().state())
+                .isEqualTo(state);
+    }
+
+    @Test
+    @SuppressWarnings("deprecation")
+    @SneakyThrows
+    void get_simplified_should_return_latest_state_in_collected_variable(){
+        //GIVEN
+        List<SurveyUnitDocument> surveyUnitDocuments = new ArrayList<>();
+
+        SurveyUnitDocument surveyUnitDocument = new SurveyUnitDocument();
+        surveyUnitDocument.setMode(String.valueOf(Mode.WEB));
+        surveyUnitDocument.setState(DataState.COLLECTED.toString());
+        surveyUnitDocument.setRecordDate(LocalDateTime.now().minusMinutes(1));
+        surveyUnitDocument.setCollectedVariables(new ArrayList<>());
+        surveyUnitDocument.getCollectedVariables().add(new VariableDocument());
+        surveyUnitDocument.getCollectedVariables().getFirst().setVarId("VAR1");
+        surveyUnitDocument.getCollectedVariables().getFirst().setIteration(1);
+        surveyUnitDocument.getCollectedVariables().getFirst().setValue("VAR1");
+        surveyUnitDocuments.add(surveyUnitDocument);
+
+        surveyUnitDocument = new SurveyUnitDocument();
+        surveyUnitDocument.setMode(String.valueOf(Mode.WEB));
+        surveyUnitDocument.setState(DataState.EDITED.toString());
+        surveyUnitDocument.setRecordDate(LocalDateTime.now());
+        surveyUnitDocument.setCollectedVariables(new ArrayList<>());
+        surveyUnitDocument.getCollectedVariables().add(new VariableDocument());
+        surveyUnitDocument.getCollectedVariables().getFirst().setVarId("VAR1");
+        surveyUnitDocument.getCollectedVariables().getFirst().setIteration(1);
+        surveyUnitDocument.getCollectedVariables().getFirst().setValue("VAR1EDITED");
+        surveyUnitDocuments.add(surveyUnitDocument);
+
+        doReturn(SurveyUnitDocumentMapper.INSTANCE.listDocumentToListModel(surveyUnitDocuments))
+                .when(surveyUnitPersistencePortStub).findByIds(any(), any());
+
+        //WHEN
+        SurveyUnitSimplifiedDto surveyUnitSimplifiedDto = surveyUnitService.findSimplifiedByCollectionInstrumentIdAndInterrogationId(
+                "test",
+                "testInterrogation",
+                Mode.WEB
+        );
+
+        //THEN
+        Assertions.assertThat(surveyUnitSimplifiedDto).isNotNull();
+        Assertions.assertThat(surveyUnitSimplifiedDto.getVariablesUpdate()).isNotNull().hasSize(1);
+        Assertions.assertThat(surveyUnitSimplifiedDto.getVariablesUpdate().getFirst()).isNotNull();
+        Assertions.assertThat(surveyUnitSimplifiedDto.getVariablesUpdate().getFirst().state()).isEqualTo(DataState.EDITED);
+    }
+
+    @Test
+    @SuppressWarnings("deprecation")
+    @SneakyThrows
+    void get_simplified_should_return_latest_state_in_external_variable(){
+        //GIVEN
+        List<SurveyUnitDocument> surveyUnitDocuments = new ArrayList<>();
+
+        SurveyUnitDocument surveyUnitDocument = new SurveyUnitDocument();
+        surveyUnitDocument.setMode(String.valueOf(Mode.WEB));
+        surveyUnitDocument.setState(DataState.COLLECTED.toString());
+        surveyUnitDocument.setRecordDate(LocalDateTime.now().minusMinutes(1));
+        surveyUnitDocument.setExternalVariables(new ArrayList<>());
+        surveyUnitDocument.getExternalVariables().add(new VariableDocument());
+        surveyUnitDocument.getExternalVariables().getFirst().setVarId("VAR1");
+        surveyUnitDocument.getExternalVariables().getFirst().setIteration(1);
+        surveyUnitDocument.getExternalVariables().getFirst().setValue("VAR1");
+        surveyUnitDocuments.add(surveyUnitDocument);
+
+        surveyUnitDocument = new SurveyUnitDocument();
+        surveyUnitDocument.setMode(String.valueOf(Mode.WEB));
+        surveyUnitDocument.setState(DataState.FORCED.toString());
+        surveyUnitDocument.setRecordDate(LocalDateTime.now());
+        surveyUnitDocument.setExternalVariables(new ArrayList<>());
+        surveyUnitDocument.getExternalVariables().add(new VariableDocument());
+        surveyUnitDocument.getExternalVariables().getFirst().setVarId("VAR1");
+        surveyUnitDocument.getExternalVariables().getFirst().setIteration(1);
+        surveyUnitDocument.getExternalVariables().getFirst().setValue("VAR1FORCED");
+        surveyUnitDocuments.add(surveyUnitDocument);
+
+        doReturn(SurveyUnitDocumentMapper.INSTANCE.listDocumentToListModel(surveyUnitDocuments))
+                .when(surveyUnitPersistencePortStub).findByIds(any(), any());
+
+        //WHEN
+        SurveyUnitSimplifiedDto surveyUnitSimplifiedDto = surveyUnitService.findSimplifiedByCollectionInstrumentIdAndInterrogationId(
+                "test",
+                "testInterrogation",
+                Mode.WEB
+        );
+
+        //THEN
+        Assertions.assertThat(surveyUnitSimplifiedDto).isNotNull();
+        Assertions.assertThat(surveyUnitSimplifiedDto.getExternalVariables()).isNotNull().hasSize(1);
+        Assertions.assertThat(surveyUnitSimplifiedDto.getExternalVariables().getFirst()).isNotNull();
+        Assertions.assertThat(surveyUnitSimplifiedDto.getExternalVariables().getFirst().state()).isEqualTo(DataState.FORCED);
     }